+++ inst/extdata/script1.R
@@ -1,0 +1,11 @@
+add_one <- function(x) {
+  return(x + 1)
+}
+
+capitalize <- function(txt) {
+  toupper(substr(txt, 1, 1))
+}
+
+say_hello <- function(name) {
+  paste('Hello', name)
+}